LPS-204580 No need to expose `getModuleAndSymbolNames`
