+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/index.es.js
@@ -84,7 +84,6 @@ export {default as createResourceURL} from './liferay/util/portlet_url/create_re
 
 // Client Extensions API
 
-export {getModuleAndSymbolNames} from './utils/client_extensions/getModuleAndSymbolNames';
 export {default as loadClientExtensions} from './utils/client_extensions/loadClientExtensions';
 export {loadModule} from './utils/client_extensions/loadModule';
