--- modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/utils/client_extensions/getModuleAndSymbolNames.ts
@@ -1,19 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-export function getModuleAndSymbolNames(
-	importDeclaration: string
-): [string, string] {
-	const parts = importDeclaration.split(' from ');
-
-	const moduleName = parts[1].trim();
-	let symbolName = parts[0].trim();
-
-	if (symbolName.startsWith('{') && symbolName.endsWith('}')) {
-		symbolName = symbolName.substring(1, symbolName.length - 1).trim();
-	}
-
-	return [moduleName, symbolName];
-}
